+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Always compare against the same time range in comparisons with "Today"
 - Added vertical indicator line to graph to make it easier to see what's hovered / selected
 - Added comparison support to funnels
+- Added revenue and revenue per visitor to funnel steps with a revenue goal
 
 ### Removed